Refactoring Errors object
